+++ src/main/java/cn/stylefeng/guns/modular/system/warpper/DeptWarpper.java
@@ -50,7 +50,7 @@ public class DeptWarpper extends BaseControllerWrapper {
 
     @Override
     protected void wrapTheMap(Map<String, Object> map) {
-        Integer pid = (Integer) map.get("pid");
+        Long pid = (Long) map.get("pid");
 
         if (ToolUtil.isEmpty(pid) || pid.equals(0)) {
             map.put("pName", "--");
